+++ clients/spot/src/websocket-api/types/user-data-stream-unsubscribe-response.ts
@@ -1,5 +1,4 @@
 /* tslint:disable */
-/* eslint-disable */
 
 /**
  * Binance Spot WebSocket API
@@ -18,10 +17,6 @@
  * https://openapi-generator.tech
  * Do not edit the class manually.
  */
-
-// May contain unused imports in some cases
-// @ts-ignore
-import type { UserDataStreamSubscribeSignatureResponseResult } from './user-data-stream-subscribe-signature-response-result';
 
 /**
  *
@@ -43,8 +38,8 @@ export interface UserDataStreamUnsubscribeResponse {
     status?: number;
     /**
      *
-     * @type {UserDataStreamSubscribeSignatureResponseResult}
+     * @type {object}
      * @memberof UserDataStreamUnsubscribeResponse
      */
-    result?: UserDataStreamSubscribeSignatureResponseResult;
+    result?: object;
 }